move datetime utils to localstack.utils.time
